Casting height/width system variables to float.

+++ Viewport.pde
@@ -4,18 +4,18 @@ class Viewport {
   float left = 0.f, right = 0.f, top = 0.f, bottom = -0.f;
   
   float getPixelPositionX(float viewportX) {
-    return ((viewportX - left) / (right - left)) * width;
+    return ((viewportX - left) / (right - left)) * (float)width;
   }
   
   float getPixelPositionY(float viewportY) {
-    return height - (((viewportY - bottom) / (top - bottom)) * height);
+    return (float)height - (((viewportY - bottom) / (top - bottom)) * (float)height);
   }
  
   float unitWidth() {
-    return width / (right - left);
+    return (float)width / (right - left);
   }
   
   float unitHeight() {
-    return height / (top - bottom);
+    return (float)height / (top - bottom);
   }
 }
